Made an extra field in the taskrequests_step-model: executer
This allows to overrule the executer in the step-model/table

Whenever a step progresses, it's now checked if for this step the superior
needs to execute it. If so: then taskrequests_step.executer is filled in
and saved: this greatly reduces the complexity of finding out the todo of
the current user.

Todo: if user is superior of own osn, then user is returned, which is wrong:
When user is superior of own osn, the superior of higher osn should be returned